+++ persistence-modules/java-jpa-2/src/main/java/com/baeldung/jpa/generateidvalue/IdGenerator.java
@@ -1,4 +1,4 @@
-package com.baeldung.jpa.primarykey;
+package com.baeldung.jpa.generateidvalue;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
