+++ persistence-modules/java-jpa-2/src/main/java/com/baeldung/jpa/generateidvalue/Task.java
@@ -1,4 +1,4 @@
-package com.baeldung.jpa.primarykey;
+package com.baeldung.jpa.generateidvalue;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -6,16 +6,19 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import javax.persistence.TableGenerator;
 
 @Entity
-@Table(name = "app_user")
-public class User {
-
+@Table(name = "task")
+public class Task {
+    
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @TableGenerator(name = "id_generator", table = "id_gen", pkColumnName = "gen_name", 
+        valueColumnName = "gen_value", pkColumnValue="task_gen", initialValue=10000, allocationSize=10)
+    @GeneratedValue(strategy = GenerationType.TABLE, generator = "id_generator")
     private Long id;
 
-    @Column(name = "user_name")
+    @Column(name = "name")
     private String name;
 
     public Long getId() {
@@ -33,5 +36,5 @@ public class User {
     public void setName(String name) {
         this.name = name;
     }
-    
+
 }
